+++ src/main/java/com/anthonyhilyard/merchantmarkers/Loader.java
@@ -5,11 +5,9 @@ import org.apache.logging.log4j.Logger;
 
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.fml.loading.FMLEnvironment;
-import net.minecraftforge.fml.config.ModConfig;
 
 @Mod(Loader.MODID)
 public class Loader
@@ -21,15 +19,13 @@ public class Loader
 	{
 		if (FMLEnvironment.dist == Dist.CLIENT)
 		{
+			if (!MerchantMarkersConfig.register(MerchantMarkersConfig.class, MODID))
+			{
+				LOGGER.error("Failed to register configuration for Merchant Markers!");
+			}
 			MerchantMarkers mod = new MerchantMarkers();
 			FMLJavaModLoadingContext.get().getModEventBus().addListener(mod::onClientSetup);
 			MinecraftForge.EVENT_BUS.register(MerchantMarkers.class);
-
-			ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, MerchantMarkersConfig.SPEC);
-		}
-		else
-		{
-			LOGGER.error("Running on a dedicated server, disabling mod.");
 		}
 	}
 }